+++ src/main/java/com/LetMeDoWith/LetMeDoWith/application/auth/service/CreateTokenService.java
@@ -2,16 +2,16 @@ package com.LetMeDoWith.LetMeDoWith.application.auth.service;
 
 import com.LetMeDoWith.LetMeDoWith.application.auth.dto.CreateRefreshTokenResult;
 import com.LetMeDoWith.LetMeDoWith.application.auth.dto.CreateTokenResult;
-import com.LetMeDoWith.LetMeDoWith.application.auth.provider.AccessTokenProvider;
-import com.LetMeDoWith.LetMeDoWith.application.auth.provider.OidcIdTokenProvider;
-import com.LetMeDoWith.LetMeDoWith.application.auth.provider.RefreshTokenProvider;
-import com.LetMeDoWith.LetMeDoWith.application.auth.provider.SignupTokenProvider;
-import com.LetMeDoWith.LetMeDoWith.application.auth.repository.RefreshTokenRepository;
+import com.LetMeDoWith.LetMeDoWith.common.auth.provider.AccessTokenProvider;
+import com.LetMeDoWith.LetMeDoWith.common.auth.provider.OidcIdTokenProvider;
+import com.LetMeDoWith.LetMeDoWith.common.auth.provider.RefreshTokenProvider;
+import com.LetMeDoWith.LetMeDoWith.common.auth.provider.SignupTokenProvider;
 import com.LetMeDoWith.LetMeDoWith.common.exception.RestApiException;
 import com.LetMeDoWith.LetMeDoWith.common.util.HeaderUtil;
 import com.LetMeDoWith.LetMeDoWith.domain.auth.enums.SocialProvider;
 import com.LetMeDoWith.LetMeDoWith.domain.auth.model.AccessToken;
 import com.LetMeDoWith.LetMeDoWith.domain.auth.model.RefreshToken;
+import com.LetMeDoWith.LetMeDoWith.domain.auth.repository.RefreshTokenRepository;
 import com.LetMeDoWith.LetMeDoWith.domain.member.model.Member;
 import com.LetMeDoWith.LetMeDoWith.domain.member.repository.MemberRepository;
 import com.LetMeDoWith.LetMeDoWith.domain.member.service.SocialAuthMemberService;
