+++ src/main/java/com/LetMeDoWith/LetMeDoWith/common/auth/provider/RefreshTokenProvider.java
@@ -1,9 +1,9 @@
-package com.LetMeDoWith.LetMeDoWith.application.auth.provider;
+package com.LetMeDoWith.LetMeDoWith.common.auth.provider;
 
-import com.LetMeDoWith.LetMeDoWith.application.auth.repository.RefreshTokenRepository;
 import com.LetMeDoWith.LetMeDoWith.common.exception.RestApiException;
 import com.LetMeDoWith.LetMeDoWith.common.exception.status.FailResponseStatus;
 import com.LetMeDoWith.LetMeDoWith.domain.auth.model.RefreshToken;
+import com.LetMeDoWith.LetMeDoWith.domain.auth.repository.RefreshTokenRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -19,14 +19,7 @@ public class RefreshTokenProvider {
     @Value("${auth.jwt.rtk-duration-day}")
     private Long rtkDurationDay;
 
-    /**
-     * 서버 Refresh Token 생성
-     *
-     * @param memberId
-     * @param accessToken
-     * @param userAgent
-     * @return
-     */
+    /** 서버 Refresh Token 생성 */
     public RefreshToken generateToken(String memberId, String accessToken, String userAgent) {
         // redis에 저장
         return refreshTokenRepository.save(
