+++ src/main/java/com/LetMeDoWith/LetMeDoWith/domain/member/vo/MemberAlarmSettingVO.java
@@ -1,0 +1,6 @@
+package com.LetMeDoWith.LetMeDoWith.domain.member.vo;
+
+import lombok.Builder;
+
+@Builder
+public record MemberAlarmSettingVO(boolean baseAlarmYn, boolean feedbackYn, boolean todoBotYn, boolean marketingYn) {}
