+++ src/main/java/com/LetMeDoWith/LetMeDoWith/infrastructure/auth/client/AppleAuthClient.java
@@ -1,7 +1,7 @@
 package com.LetMeDoWith.LetMeDoWith.infrastructure.auth.client;
 
-import com.LetMeDoWith.LetMeDoWith.application.auth.client.AuthClient;
-import com.LetMeDoWith.LetMeDoWith.application.auth.dto.OidcPublicKeyResDto;
+import com.LetMeDoWith.LetMeDoWith.common.auth.client.AuthClient;
+import com.LetMeDoWith.LetMeDoWith.common.auth.dto.OidcPublicKeyResDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.context.annotation.Profile;
