+++ src/main/java/com/LetMeDoWith/LetMeDoWith/infrastructure/auth/client/DevKakaoAuthClient.java
@@ -1,7 +1,7 @@
 package com.LetMeDoWith.LetMeDoWith.infrastructure.auth.client;
 
-import com.LetMeDoWith.LetMeDoWith.application.auth.client.AuthClient;
-import com.LetMeDoWith.LetMeDoWith.application.auth.dto.OidcPublicKeyResDto;
+import com.LetMeDoWith.LetMeDoWith.common.auth.client.AuthClient;
+import com.LetMeDoWith.LetMeDoWith.common.auth.dto.OidcPublicKeyResDto;
 import java.util.List;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
