+++ src/main/java/com/LetMeDoWith/LetMeDoWith/infrastructure/auth/client/KakaoAuthClient.java
@@ -1,7 +1,7 @@
 package com.LetMeDoWith.LetMeDoWith.infrastructure.auth.client;
 
-import com.LetMeDoWith.LetMeDoWith.application.auth.client.AuthClient;
-import com.LetMeDoWith.LetMeDoWith.application.auth.dto.OidcPublicKeyResDto;
+import com.LetMeDoWith.LetMeDoWith.common.auth.client.AuthClient;
+import com.LetMeDoWith.LetMeDoWith.common.auth.dto.OidcPublicKeyResDto;
 import com.LetMeDoWith.LetMeDoWith.common.cache.CacheName;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.CacheConfig;
